Bit of refactoring to separate the functionality of the various queueing systems (i.e. the PDF conversion queue, the scan queue, and the email queue).
Added a scanner base class from which the (for now, two) types of scanner can inherit. Then we can just have a single Queue which takes a Scanner, and calls its scan() method, regardless of whether the queued scanner is for Images, or URLs, or anything else.

Might have been a bit heavy with the classes (e.g. the data classes which just hold email addresses, or file output directories); perhaps the more Pythonic way would be to just pass around tuples or dictionaries. I'll come back to this when things are working a bit better.