Removed separate PDF converter; PDF scanner now converts a PDF and scans it in a single thread.
To mitigate disk space issues, this scanner now converts each page of a PDF, scans that page, and then deletes it, before moving onto the next page. It's a bit slower, but it does mean that a large PDF won't be converted into a load of large PNG files that will take up a lot of space.

Running this on and SSD VPS means 1) disk space is probably relatively limited, but also 2) I/O is going to be pretty fast, so hopefully the read/convert/write/delete cycle won't be too cumbersome.