If the `send_email()` method is a class method on `EmailThread`, we can only send a number of emails equal to the number of threads that are started by the `EmailManager`. So there's some sort of resource not being released.
After trying to work out why things worked for `PdfScanner` (i.e. with two threads, you can keep uploading PDFs and they would keep getting converted as usual), I noticed that the `PdfScanner` imports its `scan_pdf` method; it is not a class method of `PdfScanner`.

Moving the `send_email()` method *outside* of the `EmailThread` class seems to fix the problem, but I DON'T KNOW WHY (or if it's the right way to do it). So whilst this works (or appears to), I need to look into and try to understand what's going on. But that's the whole point of this exercise, so it's a positive (if frustrating) development.